Moving YDB headers from yql/essentials
