+++ ydb/library/yql/dq/opt/dq_opt_join.h
@@ -9,7 +9,6 @@
 namespace NYql {
 
 struct TOptimizerStatistics;
-struct TRelOptimizerNode;
 
 namespace NDq {
