+++ yql/essentials/core/cbo/ya.make
@@ -1,12 +1,12 @@
 LIBRARY()
 
-SRCS(
-    cbo_optimizer_new.cpp
-    cbo_interesting_orderings.cpp
-    cbo_hints.cpp
-)
+# Type definitions and implementations have moved to ydb/core/kqp/opt/cbo/.
+# This library is now a thin alias wrapper — no SRCS, just re-exports the ydb libraries.
 
-GENERATE_ENUM_SERIALIZATION(cbo_optimizer_new.h)
+PEERDIR(
+    ydb/core/kqp/opt/cbo
+    ydb/core/kqp/opt/cbo/optimizer
+)
 
 END()
