+++ yql/essentials/core/statistics/ya.make
@@ -1,0 +1,15 @@
+LIBRARY()
+
+SRCS(
+    yql_statistics.cpp
+)
+
+PEERDIR(
+    ydb/core/kqp/opt/cbo
+    yql/essentials/core/minsketch
+    yql/essentials/core/histogram
+    library/cpp/json
+    library/cpp/string_utils/base64
+)
+
+END()
