+++ yql/essentials/core/yql_cost_function.cpp
@@ -1,15 +1,3 @@
 
+// operator< for TJoinColumn has moved to ydb/core/kqp/opt/cbo/cbo_interesting_orderings.cpp
 #include "yql_cost_function.h"
-
-namespace NYql {
-
-bool NDq::operator < (const NDq::TJoinColumn& c1, const NDq::TJoinColumn& c2) {
-    if (c1.RelName < c2.RelName){
-        return true;
-    } else if (c1.RelName == c2.RelName) {
-        return c1.AttributeName < c2.AttributeName;
-    }
-    return false;
-}
-
-}  // namespace NYql
